requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
holidays
imbalanced-learn
joblib
matplotlib
numpy
optuna
pandas
pandas_market_calendars
plotly
python-dotenv
scikit-learn
seaborn
streamlit
ta
xgboost
yfinance